+++ src/brevitas_examples/llm/llm_quant/parse_utils.py
@@ -8,7 +8,7 @@ from pathlib import Path
 from brevitas.utils.logging import setup_logger
 from brevitas.utils.python_utils import Registry
 from brevitas_examples.common.generative.quantizers import QUANTIZERS_REGISTRY
-from brevitas_examples.llm.llm_quant.rotation_optimization import TRAINER_REGISTRY
+from brevitas_examples.llm.llm_quant.trainer_utils import TRAINER_REGISTRY
 
 logging = setup_logger(__name__)
